feat: add relationship and notification model
