+++ app/src/main/java/com/example/nocket/ui/screen/home/HomeScreen.kt
@@ -2,16 +2,8 @@ package com.example.nocket.ui.screen.home
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.Message
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -19,11 +11,9 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
-import com.example.nocket.Screen
-import com.example.nocket.components.common.MyTopBar
 import com.example.nocket.components.grid.PostGrid
+import com.example.nocket.components.topbar.MainTopBar
 import com.example.nocket.data.SampleData
 import com.example.nocket.models.Post
 import com.example.nocket.ui.screen.post.CameraScreen
@@ -43,16 +33,18 @@ fun HomeScreen(
                 onBack = { selectedPost = null }
             )
         }
+
         showCameraMode -> {
             CameraScreen(
                 onBack = { showCameraMode = false },
                 onPhotoTaken = { showCameraMode = false }
             )
         }
+
         else -> {
             Scaffold(
                 topBar = {
-                    MyTopBar(
+                    MainTopBar(
                         navController = navController,
                         user = SampleData.users[0],
                     )
